+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.2
+ - Change import button to Person+
+ - i18n button names on import dialog
+ - Add ability to specify additional compendiums or re-order them
+ - Add race compendium - not used yet
+
 ## 0.6.1
 - MPMB: Correctly handle multi-classing
 - MPMB: Item importing: match if you can, or import the name into the Loot category